Ongoing - refactoring charts, adding new tab/section
